requirements.txt
coverage==4.2
Django>=2.0,<2.1
django-cors-headers==2.0.2
django-debug-toolbar==1.9.1
django-extensions==2.1.3
django-filter==0.13.0
djangorestframework>=3.8,<3.9
djoser==0.5.0
ipdb==0.10.1
ipython==7.0.1
Markdown==2.6.6
Werkzeug==0.14.1
psycopg2==2.7.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
boto3==1.4.4
botocore==1.5.80
